refactor console state
