+++ apps/src/javalab/Javalab.js
@@ -9,22 +9,24 @@ import javalab, {
   setAllSourcesAndFileMetadata,
   setAllValidation,
   setDisplayTheme,
-  appendOutputLog,
   setIsStartMode,
   setLevelName,
-  appendNewlineToConsoleLog,
   setIsRunning,
   setIsTesting,
-  openPhotoPrompter,
-  closePhotoPrompter,
   setBackpackEnabled,
-  appendMarkdownLog,
   setIsReadOnlyWorkspace,
   setHasOpenCodeReview,
   setValidationPassed,
   setHasRunOrTestedCode,
   setIsJavabuilderConnecting
 } from './javalabRedux';
+import javalabConsole, {
+  appendOutputLog,
+  appendNewlineToConsoleLog,
+  appendMarkdownLog,
+  closePhotoPrompter,
+  openPhotoPrompter
+} from './redux/consoleRedux';
 import {TestResults} from '@cdo/apps/constants';
 import project from '@cdo/apps/code-studio/initApp/project';
 import JavabuilderConnection from './JavabuilderConnection';
@@ -191,7 +193,7 @@ Javalab.prototype.init = function(config) {
     isSubmitted: !!config.level.submitted
   });
 
-  registerReducers({javalab});
+  registerReducers({javalab, javalabConsole});
   // If we're in editBlock mode (for editing start_sources) we set up the save button to save
   // the project file information into start_sources on the level.
   if (this.isStartMode) {
